refactor: extract commit into core module
