+++ src/gcp.rs
@@ -1,14 +1,14 @@
 use crate::DatastoreEmulator;
-use crate::database::{DatastoreStorage, EntityWithMetadata, KeyStruct, TransactionState};
+use crate::database::{DatastoreStorage, TransactionState};
 use crate::google::datastore::v1::aggregation_query::aggregation::Operator as AggregationOperator;
 use crate::google::datastore::v1::datastore_server::Datastore as DatastoreService;
 use crate::google::datastore::v1::{
     AggregationResultBatch, AllocateIdsRequest, AllocateIdsResponse, BeginTransactionRequest,
-    BeginTransactionResponse, CommitRequest, CommitResponse, Entity, EntityResult, ExecutionStats,
-    ExplainMetrics, Filter, LookupRequest, LookupResponse, PingRequest, PingResponse, PlanSummary,
+    BeginTransactionResponse, CommitRequest, CommitResponse, ExecutionStats, ExplainMetrics,
+    Filter, LookupRequest, LookupResponse, PingRequest, PingResponse, PlanSummary,
     PropertyReference, ReserveIdsRequest, ReserveIdsResponse, RollbackRequest, RollbackResponse,
     RunAggregationQueryRequest, RunAggregationQueryResponse, RunQueryRequest, RunQueryResponse,
-    commit_request::TransactionSelector, key::path_element::IdType, mutation::Operation,
+    key::path_element::IdType,
 };
 use pbjson_types::value::Kind;
 use pbjson_types::{Duration, Struct, Value as ValueProps};
@@ -71,278 +71,8 @@ impl DatastoreService for DatastoreEmulator {
         &self,
         request: Request<CommitRequest>,
     ) -> Result<Response<CommitResponse>, Status> {
-        tracing::debug!("Received CommitRequest: {:?}", request);
-        let req = request.into_inner();
-        let mut storage = self.storage.write().await;
-        let mut mutation_results = Vec::new();
-
-        // Handle transaction if present
-        let transaction_id = if let Some(transaction_selector) = req.transaction_selector {
-            match transaction_selector {
-                TransactionSelector::Transaction(tx_bytes) => {
-                    // Convert transaction ID bytes to string
-                    match String::from_utf8(tx_bytes.clone()) {
-                        Ok(id) => Some(id),
-                        Err(_) => {
-                            return Err(Status::invalid_argument("Invalid transaction ID format"));
-                        }
-                    }
-                }
-                TransactionSelector::SingleUseTransaction(_) => None, // No existing transaction, just commit directly
-            }
-        } else {
-            None // No transaction specified, just commit directly
-        };
-
-        // Check if this is a transactional commit
-        if let Some(tx_id) = transaction_id.clone() {
-            // Verify the transaction exists and is not read-only
-            if let Some(_tx_state) = storage.transactions.get(&tx_id) {
-                // if tx_state.read_only {
-                //     return Err(Status::failed_precondition(
-                //         "Cannot commit mutations in a read-only transaction blabla",
-                //     ));
-                // }
-
-                // In a real implementation, we would apply the transaction's mutations here
-                // For now, we'll just remove the transaction after committing
-                tracing::info!("Committing transaction: {}", tx_id);
-            } else {
-                return Err(Status::not_found(format!(
-                    "Transaction {} not found",
-                    tx_id
-                )));
-            }
-        }
-
-        // Process mutations
-        for mutation in req.mutations {
-            if let Some(ref mutation) = mutation.operation {
-                match mutation {
-                    Operation::Insert(entity) => {
-                        //dbg!("Attempting to insert entity (from main.rs)", &entity);
-                        match storage.insert_entity(entity) {
-                            Ok((final_key, metadata)) => {
-                                mutation_results.push(
-                                    crate::google::datastore::v1::MutationResult {
-                                        key: Some(final_key),
-                                        version: metadata.version as i64,
-                                        create_time: Some(metadata.create_time.clone()),
-                                        update_time: Some(metadata.update_time.clone()),
-                                        conflict_detected: false,
-                                        transform_results: vec![],
-                                    },
-                                );
-                            }
-                            Err(status) => {
-                                return Err(status);
-                            }
-                        }
-                    }
-                    Operation::Update(entity) => {
-                        //dbg!("Updating entity", &entity);
-                        let key = match entity.key {
-                            Some(ref key) => key.clone(),
-                            None => {
-                                return Err(Status::invalid_argument(
-                                    "Entity missing key for update",
-                                ));
-                            }
-                        };
-                        let key_struct = KeyStruct::from_datastore_key(&key);
-
-                        // Option to store data extracted from the mutable borrow if entity is found
-                        let mut opt_updated_data: Option<(
-                            /*entity_for_index_update*/ Entity,
-                            /*version_for_result*/ u64,
-                            /*create_time_for_result*/ pbjson_types::Timestamp,
-                            /*update_time_for_result*/ pbjson_types::Timestamp,
-                        )> = None;
-
-                        if let Some(existing_entity_metadata) =
-                            storage.entities.get_mut(&key_struct)
-                        {
-                            if let Some(_k) = &existing_entity_metadata.entity.key {
-                                //dbg!("Updating entity", &k.path);
-                            }
-
-                            let timestamp_now = pbjson_types::Timestamp {
-                                // Placeholder
-                                seconds: 0,
-                                nanos: 0,
-                            };
-
-                            existing_entity_metadata.entity = entity.clone(); // `entity` is from the request
-                            existing_entity_metadata.version += 1;
-                            existing_entity_metadata.update_time = timestamp_now.clone();
-
-                            // Store all data needed after this block, then the borrow of existing_entity_metadata ends
-                            opt_updated_data = Some((
-                                existing_entity_metadata.entity.clone(), // For update_indexes
-                                existing_entity_metadata.version,        // For MutationResult
-                                existing_entity_metadata.create_time.clone(), // For MutationResult
-                                timestamp_now, // For MutationResult (it's existing_entity_metadata.update_time)
-                            ));
-                        } // Mutable borrow of storage.entities (existing_entity_metadata) ends here.
-
-                        // Now, call storage.update_indexes and push to mutation_results if data was updated
-                        if let Some((entity_for_index, version, create_time, update_time)) =
-                            opt_updated_data
-                        {
-                            storage.update_indexes(&key_struct, &entity_for_index);
-
-                            mutation_results.push(crate::google::datastore::v1::MutationResult {
-                                key: entity.key.clone(), // Key from the original request entity
-                                version: version as i64,
-                                create_time: Some(create_time),
-                                update_time: Some(update_time),
-                                conflict_detected: false,
-                                transform_results: vec![],
-                            });
-                        }
-                    } // Closes Operation::Update(entity) arm
-                    // Removed duplicate Operation::Update block
-                    Operation::Upsert(entity) => {
-                        let key = match entity.key {
-                            Some(ref key) => key.clone(),
-                            None => {
-                                return Err(Status::invalid_argument(
-                                    "Entity missing key for upsert",
-                                ));
-                            }
-                        };
-
-                        let key_is_incomplete = key
-                            .path
-                            .iter()
-                            .any(|path_element| path_element.id_type.is_none());
-
-                        if key_is_incomplete {
-                            let (final_key, metadata) = storage.insert_entity(&entity)?;
-                            mutation_results.push(crate::google::datastore::v1::MutationResult {
-                                key: Some(final_key),
-                                version: metadata.version as i64,
-                                create_time: Some(metadata.create_time.clone()),
-                                update_time: Some(metadata.update_time.clone()),
-                                conflict_detected: false,
-                                transform_results: vec![],
-                            });
-                        } else {
-                            let key_struct = KeyStruct::from_datastore_key(&key);
-
-                            let timestamp_now = pbjson_types::Timestamp {
-                                // Placeholder
-                                seconds: 0,
-                                nanos: 0,
-                            };
-
-                            let mut observe_key_counters = false;
-                            let entry = storage.entities.entry(key_struct.clone());
-                            let version;
-                            let create_time;
-                            let update_time = timestamp_now.clone();
-
-                            match entry {
-                                std::collections::btree_map::Entry::Occupied(
-                                    mut occupied_entry,
-                                ) => {
-                                    let metadata = occupied_entry.get_mut();
-                                    metadata.entity = entity.clone();
-                                    metadata.version += 1;
-                                    metadata.update_time = update_time.clone();
-                                    version = metadata.version;
-                                    create_time = metadata.create_time.clone();
-                                }
-                                std::collections::btree_map::Entry::Vacant(vacant_entry) => {
-                                    let new_metadata = EntityWithMetadata {
-                                        entity: entity.clone(),
-                                        version: 1,
-                                        create_time: timestamp_now.clone(),
-                                        update_time: update_time.clone(),
-                                    };
-                                    version = new_metadata.version;
-                                    create_time = new_metadata.create_time.clone();
-                                    observe_key_counters = true;
-                                    vacant_entry.insert(new_metadata);
-                                }
-                            }
-
-                            if observe_key_counters {
-                                storage.observe_key_id(&key);
-                            }
-
-                            storage.update_indexes(&key_struct, &entity);
-
-                            mutation_results.push(crate::google::datastore::v1::MutationResult {
-                                key: Some(key),
-                                version: version as i64,
-                                create_time: Some(create_time.clone()),
-                                update_time: Some(update_time.clone()),
-                                conflict_detected: false,
-                                transform_results: vec![],
-                            });
-                        }
-                    }
-                    Operation::Delete(key_to_delete) => {
-                        if let Some(removed_entity_metadata) = storage.delete_entity(key_to_delete)
-                        {
-                            if let Some(_k) = &removed_entity_metadata.entity.key {
-                                //dbg!("Deleting entity", &k.path);
-                            }
-
-                            let timestamp_now = pbjson_types::Timestamp {
-                                // Placeholder
-                                seconds: 0,
-                                nanos: 0,
-                            };
-                            mutation_results.push(crate::google::datastore::v1::MutationResult {
-                                key: Some(key_to_delete.clone()), // Return the original key from request
-                                version: removed_entity_metadata.version as i64, // Use actual version
-                                create_time: Some(removed_entity_metadata.create_time.clone()),
-                                update_time: Some(timestamp_now), // Deletion time could be now
-                                conflict_detected: false,
-                                transform_results: vec![],
-                            });
-                        } else {
-                            tracing::warn!(
-                                "Entity not found for deletion with key: {:?}",
-                                key_to_delete.path
-                            );
-                            // Datastore typically doesn't error if deleting a non-existent entity.
-                            // It returns a MutationResult with no key or version, or an empty result list.
-                            // For simplicity, we can add a result indicating nothing happened or skip.
-                            // Adding a result with the key but perhaps version 0 or specific times.
-                            mutation_results.push(crate::google::datastore::v1::MutationResult {
-                                key: Some(key_to_delete.clone()),
-                                version: 0, // Indicate not found or already deleted
-                                create_time: None,
-                                update_time: None,
-                                conflict_detected: false,
-                                transform_results: vec![],
-                            });
-                        }
-                    }
-                }
-            } else {
-                //  What should return if there is not mutation?
-            }
-        }
-
-        let index_updates = mutation_results.len() as i32;
-        // Clean up the transaction if it was used
-        if let Some(tx_id) = transaction_id {
-            storage.clean_transaction(&tx_id);
-        }
-
-        // Get current time for the commit timestamp
-        let commit_time = system_time_to_timestamp(SystemTime::now());
-
-        let response = Response::new(CommitResponse {
-            mutation_results,
-            index_updates,
-            commit_time: Some(commit_time),
-        });
-        Ok(response)
+        let resp = crate::core::commit(&self.storage, request.into_inner()).await?;
+        Ok(Response::new(resp))
     }
 
     async fn begin_transaction(
